% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-utils.R
\name{add_cohort_name}
\alias{add_cohort_name}
\title{Add Cohort Name}
\usage{
add_cohort_name(input_data, cohort_name, cohort_name_column = "cohort_name")
}
\arguments{
\item{input_data}{The input data as a
data frame or tibble.}

\item{cohort_name}{A text indicating the name of
the cohort}

\item{cohort_name_column}{A text indicating the
column name used to represent the name of the cohort.
Default: 'cohort_name'}
}
\value{
A data frame with a new column called
the input text from`cohort_name_column` containing
a character vector of the input text from `cohort_name`.
}
\description{
Add cohort name as a new column
in the dataset and move it to the first column.
}
\details{
An error will be raised if input
`cohort_name_column` is already a column name
in `input_data`. Currently, no overriding of
columns is allowed.
}
\examples{

input_data <- tibble::tribble(
  ~column_a, ~column_b,
  1, "Yes",
  1, "No")

cohort_name = "Cohort 1"

result_data <- input_data |>
  add_cohort_name(cohort_name = "Cohort 1",
                  cohort_name_column = "cohorts")

result_data

}
