% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputdata-excel.R
\name{calculate_column_max_char}
\alias{calculate_column_max_char}
\title{Calculate Column Maximum Character}
\usage{
calculate_column_max_char(data)
}
\arguments{
\item{data}{Input tibble or data frame.}
}
\value{
A numeric vector each value indicated the
maximum number of characters for each column of the
input data frame or tibble.
}
\description{
Calculate the maximum number of
characters for each column in the input data frame
or tibble, including the column name.
}
\examples{
coronary_vessel_segments <- c(
  "LM", "pLAD", "mLAD", "dLAD",
  "pLCX", "dLCX", "pRCA", "mRCA", "dRCA"
)

vessel_severity <- c(
  "Severe", "Severe", "Minimal", "Occluded",
  "Moderate", "Moderate", "Normal", "Severe", "Mild"
)


vessel_disease_data <- data.frame(
  coronary_vessel_segments = coronary_vessel_segments,
  vessel_severity = vessel_severity
)

column_max_char <- calculate_column_max_char(vessel_disease_data)

column_max_char

}
