% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-reactable.R
\name{create_dropdown_filter_render_function}
\alias{create_dropdown_filter_render_function}
\title{Create Drop Down Filter Render Function}
\usage{
create_dropdown_filter_render_function(
  reactable_element_id,
  style = "width: 100\%; height: 100\%;"
)
}
\arguments{
\item{reactable_element_id}{Text input indicating the
element ID of the `reactable`}

\item{style}{Text input indicating the javascript style parameters
of the filter input drop down box.
Default: 'width: 100\%; height: 28px;'}
}
\value{
A render function for a `reactable` column indicating that
it needs a drop down filter input box.
}
\description{
Function that returns a `reactable` filter input
render function that provide a drop down functionality.
}
\examples{
create_dropdown_filter_render_function(
  reactable_element_id = "drop-down-filter-table")
}
