% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{get_knitr_version}
\alias{get_knitr_version}
\title{Get `knitr` Version}
\usage{
get_knitr_version()
}
\value{
A text saying the version number
and where it is downloaded from.
}
\description{
A function that gives some
information on the loaded library `knitr`.
}
\examples{
library("knitr")

knitr_info <- get_knitr_version()

knitr_info

}
