% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stenosis.R
\name{get_max_stenosis_severity}
\alias{get_max_stenosis_severity}
\title{Get Max Stenosis Severity}
\usage{
get_max_stenosis_severity(vessel_data, stenosis_severity_grp_column)
}
\arguments{
\item{vessel_data}{Input coronary vessel data}

\item{stenosis_severity_grp_column}{Column in
`vessel_data` indicating the severity level of the
different coronary vessels. Accepted severity levels are
\itemize{
  \item Normal (0\% stenosis)
  \item Minimal (1 to 24\% stenosis)
  \item Mild (25 to 49\% stenosis)
  \item Moderate (50 to 69\% stenosis)
  \item Severe (70 to 99\% stenosis)
  \item Occluded (100\% stenosis)
}}
}
\value{
A text indicating the highest level of stenosis
severity from coronary vessel data.
}
\description{
Obtain the highest level of stenosis
severity from coronary vessel data.
}
\details{
From
\href{https://doi.org/10.1016/j.jcct.2022.07.002}{Cury et. al. (2022)},
the Society of Cardiovascular Computed Tomography (SCCT)
graded luminal stenosis as follows:
\itemize{
  \item Normal (0\% stenosis)
  \item Minimal (1 to 24\% stenosis)
  \item Mild (25 to 49\% stenosis)
  \item Moderate (50 to 69\% stenosis)
  \item Severe (70 to 99\% stenosis)
  \item Occluded (100\% stenosis)
}
}
\examples{
occluded_coronary_vessel_data <- tibble::tribble(
  ~coronary_vessel, ~vessel_severity,
  "LM", "Normal",
  "LAD", "Severe",
  "LCX", "Minimal",
  "RCA", "Occluded"
)

max_stenosis_severity <- get_max_stenosis_severity(
  occluded_coronary_vessel_data,
  "vessel_severity"
)

max_stenosis_severity
}
