% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stenosis.R
\name{get_number_of_vessels_with_high_stenosis_severity_exclude_LM}
\alias{get_number_of_vessels_with_high_stenosis_severity_exclude_LM}
\title{Get Number Of Vessels With High Stenosis Severity (Excluding LM)}
\usage{
get_number_of_vessels_with_high_stenosis_severity_exclude_LM(
  vessel_data,
  vessel_name_column,
  stenosis_severity_grp_column,
  lad_vessel_segments_grp_name = "LAD",
  lcx_vessel_segments_grp_name = "LCX",
  rca_vessel_segments_grp_name = "RCA",
  lad_vessel_segments = c("pLAD", "mLAD", "dLAD", "Diag1", "Diag2", "Diag3", "Diag4",
    "Ramus1", "Ramus2"),
  lcx_vessel_segments = c("pLCX", "dLCX", "LPDA", "LPL", "OM1", "OM2", "OM3", "OM4",
    "OM5"),
  rca_vessel_segments = c("pRCA", "mRCA", "dRCA", "RPDA", "RPL")
)
}
\arguments{
\item{vessel_data}{Input coronary vessel data}

\item{vessel_name_column}{Column in
`vessel_data` indicating the different vessel
segments for LM, LAD, LCX and RCA.}

\item{stenosis_severity_grp_column}{Column in
`vessel_data` indicating the severity level of the
different coronary vessels. Accepted severity levels are
\itemize{
  \item Normal (0\% stenosis)
  \item Minimal (1 to 24\% stenosis)
  \item Mild (25 to 49\% stenosis)
  \item Moderate (50 to 69\% stenosis)
  \item Severe (70 to 99\% stenosis)
  \item Occluded (100\% stenosis)
}}

\item{lad_vessel_segments_grp_name}{Name of vessel
segment group to indicate left anterior descending
or LAD vessels.
Default: 'LAD'}

\item{lcx_vessel_segments_grp_name}{Name of vessel
segment group to indicate left circumflex
or LCX vessels.
Default: 'LCX'}

\item{rca_vessel_segments_grp_name}{Name of vessel
segment group to indicate right coronary artery
or RCA.
Default: 'RCA'}

\item{lad_vessel_segments}{Name of vessel
segments in the `vessel_name_column` to be
grouped as left anterior descending
or LAD vessels.
Default: c("pLAD", "mLAD", "dLAD", "Diag1",
           "Diag2", "Diag3", "Diag4",
           "Ramus1", "Ramus2")}

\item{lcx_vessel_segments}{Name of vessel
segments in the `vessel_name_column` to be
grouped as left circumflex
or LCX vessels.
Default: c("pLCX", "dLCX",
           "LPDA", "LPL",
           "OM1", "OM2", "OM3", "OM4", "OM5")}

\item{rca_vessel_segments}{Name of vessel
segments in the `vessel_name_column` to be
grouped as right coronary artery
or RCA.
Default: c("pRCA", "mRCA", "dRCA", "RPDA", "RPL")}
}
\value{
An integer from 0 to 3
indicating the number of vessels with stenosis
severity of 50% or greater.
}
\description{
Get the number of vessels with stenosis
severity of 50% or greater. Do note that LM is not considered as a vessel.
}
\details{
From
\href{https://doi.org/10.1016/j.jcct.2022.07.002}{Cury et. al. (2022)},
the Society of Cardiovascular Computed Tomography (SCCT)
graded luminal stenosis as follows:
\itemize{
  \item Normal (0\% stenosis)
  \item Minimal (1 to 24\% stenosis)
  \item Mild (25 to 49\% stenosis)
  \item Moderate (50 to 69\% stenosis)
  \item Severe (70 to 99\% stenosis)
  \item Occluded (100\% stenosis)
}
Do note that LM is not considered as a vessel.
}
\examples{
zero_vessel_disease_data <- tibble::tribble(
  ~coronary_vessel_segments, ~vessel_severity,
  "LM", "Normal",
  "pLAD", "Normal",
  "mLAD", "Normal",
  "dLAD", "Normal",
  "Diag1", "Normal",
  "Diag2", "Normal",
  "Ramus", "Normal",
  "pLCX", "Normal",
  "dLCX", "Normal",
  "OM1", "Normal",
  "OM2", "Normal",
  "pRCA", "Normal",
  "mRCA", "Normal",
  "dRCA", "Normal",
  "RPDA", "Normal",
  "RPL", "Normal")

four_vessel_disease_data <- tibble::tribble(
  ~coronary_vessel_segments, ~vessel_severity,
  "LM", "Severe",
  "pLAD", "Severe",
  "mLAD", "Minimal",
  "dLAD", "Occluded",
  "Diag1", "Normal",
  "Diag2", "Normal",
  "Diag3", "Normal",
  "Diag4", "Normal",
  "pLCX", "Moderate",
  "dLCX", "Moderate",
  "OM1", "Normal",
  "OM2", "Normal",
  "pRCA", "Normal",
  "mRCA", "Severe",
  "dRCA", "Mild",
  "RPDA", "Normal",
  "RPL", "Normal")

 zero_disease_vessel <- get_number_of_vessels_with_high_stenosis_severity_exclude_LM(
   vessel_data = zero_vessel_disease_data,
   vessel_name_column = "coronary_vessel_segments",
   stenosis_severity_grp_column = "vessel_severity",
   lad_vessel_segments_grp_name = "LAD",
   lcx_vessel_segments_grp_name = "LCX",
   rca_vessel_segments_grp_name = "RCA",
   lad_vessel_segments = c("pLAD", "mLAD", "dLAD",
                           "Diag1", "Diag2", "Ramus"),
   lcx_vessel_segments = c("pLCX", "dLCX", "OM1", "OM2"),
   rca_vessel_segments = c("pRCA", "mRCA", "dRCA", "RPDA", "RPL")
   )

 four_disease_vessels <- get_number_of_vessels_with_high_stenosis_severity_exclude_LM(
   vessel_data = four_vessel_disease_data,
   vessel_name_column = "coronary_vessel_segments",
   stenosis_severity_grp_column = "vessel_severity",
   lad_vessel_segments_grp_name = "LAD",
   lcx_vessel_segments_grp_name = "LCX",
   rca_vessel_segments_grp_name = "RCA",
   lad_vessel_segments = c("pLAD", "mLAD", "dLAD",
                           "Diag1", "Diag2", "Ramus"),
   lcx_vessel_segments = c("pLCX", "dLCX", "OM1", "OM2"),
   rca_vessel_segments = c("pRCA", "mRCA", "dRCA", "RPDA", "RPL")
   )

 zero_disease_vessel

 four_disease_vessels

}
