% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{get_quarto_version}
\alias{get_quarto_version}
\title{Print Quarto Version}
\usage{
get_quarto_version(test_sys_path = FALSE, test_no_path = FALSE)
}
\arguments{
\item{test_sys_path}{Set to TRUE if we want to
test the program finding Quarto in a system
environment.
Default: FALSE}

\item{test_no_path}{Set to TRUE if we want to
test when a path to Quarto does not exist.
Default: FALSE}
}
\value{
A vector of text indicating the Quarto version
and where is it found in the computer.
}
\description{
A function that gives some
information on which version of Quarto
is used and its location. We can't name it
as get_quarto_version
}
\examples{

quarto_version <- get_quarto_version()

}
