% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environment.R
\name{get_r_package_info}
\alias{get_r_package_info}
\title{Get R Package Information}
\usage{
get_r_package_info()
}
\value{
A tibble with four columns `package`,
`version`, `date` and `source`.
\itemize{
  \item `package`: The name of the loaded package.
  \item `version`: The version number of the loaded package.
  \item `date`: The date of the loaded package was installed.
  \item `source`: The source where the loaded package was installed.
}
}
\description{
A function that gives some
information on the loaded libraries used
in a script or project.
}
\examples{
library("dplyr")

r_package_info <- get_r_package_info()

r_package_info

}
