% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputdata-excel.R
\name{write_to_sheet}
\alias{write_to_sheet}
\title{Write To Excel Sheet}
\usage{
write_to_sheet(sheet_name, data, workbook)
}
\arguments{
\item{sheet_name}{A vector of characters representing
sheet names for the Excel file.}

\item{data}{A list of tibble and data frame representing
the data needed to output to the Excel file.}

\item{workbook}{A Workbook object created using
`openxlsx::createWorkbook`.}
}
\description{
Write data to an excel sheet.
}
\examples{
coronary_vessel_segments <- c(
  "LM", "pLAD", "mLAD", "dLAD",
  "pLCX", "dLCX", "pRCA", "mRCA", "dRCA"
)

vessel_severity <- c(
  "Severe", "Severe", "Minimal", "Occluded",
  "Moderate", "Moderate", "Normal", "Severe", "Mild"
)

vessel_disease_data <- data.frame(
  coronary_vessel_segments = coronary_vessel_segments,
  vessel_severity = vessel_severity
)

# Create a new workbook
my_workbook <- openxlsx::createWorkbook()

sheet_name = c("vessel disease")
output_data = vessel_disease_data

write_to_sheet(sheet_name = sheet_name,
               data = output_data,
               workbook = my_workbook)

}
