% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation-package.R
\docType{package}
\name{harmonisation-package}
\alias{harmonisation}
\alias{harmonisation-package}
\title{harmonisation: Utility functions for project}
\description{
Utility functions to harmonise data for project.
}
\author{
\strong{Maintainer}: Jeremy Selva \email{jeremy1189.jjs@gmail.com} (\href{https://orcid.org/0000-0002-4498-2662}{ORCID})

}
\keyword{internal}
